Convert smart quotes to dumb quotes
Since we know have the smartypants plugin installed, we might as well
change qutoes in the source mdx files to "dumb quotes".

This commit also temporarily disables the running of proselint since
proselint complains about dumb quotes. There doesn't seem to be a way
to configure proselint on a per-directory level at the moment.

- https://github.com/amperser/proselint/issues/381
- https://github.com/amperser/proselint/issues/997