Remove mdspell and proselint
mdspell became unfeasible due to the number of examples being flagged
as false positives.

Proselint doesn't permit configuring rules on a per-directory basis,
which is a deal-breaker for me because it means that I can't disable
`typography.symbols.curly_quotes`.

- https://github.com/amperser/proselint/issues/381
- https://github.com/amperser/proselint/issues/997

The projects development seems to have halted, so I doubt we'd see this
in the near future.